utilizes useState and useEffect to render created components from pokeAPI
useEffect is used here to retrievepokemon data from the pokeAPI and work in conjuction with useState to keep track of which pokemon are to be rendered on the website. The CardsDisplay component pulls the data stored in useState and maps each pokemon to its own Card component. useEffect is also used here to ensure the API is only fetched on mount.